# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for Ticket Booking System",
        "title": "Ticket Booking System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/promocodes": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает полный список промокодов для администратора. / [EN] Returns full list of promocodes for admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-promocodes"
                ],
                "summary": "[RU] Список всех промокодов / [EN] All promocodes list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "[RU] Страница / [EN] Page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "[RU] Лимит / [EN] Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "[RU] Активен (true/false) / [EN] Is active (true/false)",
                        "name": "is_active",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/promocodes/bulk-deactivate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Деактивация сразу нескольких промокодов по их ID. / [EN] Deactivation of several promocodes by their IDs at once.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-promocodes"
                ],
                "summary": "[RU] Массовая деактивация / [EN] Bulk deactivate",
                "parameters": [
                    {
                        "description": "[RU] Список ID / [EN] List of IDs",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "ids": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/promocodes/export": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Выгружает список всех промокодов в CSV. / [EN] Exports the list of all promocodes to CSV.",
                "produces": [
                    "text/csv"
                ],
                "tags": [
                    "admin-promocodes"
                ],
                "summary": "[RU] Экспорт промокодов / [EN] Export promocodes",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/admin/promocodes/stats": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Общая статистика по использованию промокодов. / [EN] General statistics on promocode usage.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-promocodes"
                ],
                "summary": "[RU] Статистика промокодов / [EN] Promocodes statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.PromocodesStats"
                        }
                    }
                }
            }
        },
        "/admin/promocodes/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Возвращает детальную информацию о промокоде, включая историю использования. / [EN] Returns detailed promocode info, including usage history.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-promocodes"
                ],
                "summary": "[RU] Детали промокода / [EN] Promocode details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID промокода / [EN] Promocode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.PromocodeDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Принудительное обновление параметров промокода. / [EN] Forced update of promocode parameters.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-promocodes"
                ],
                "summary": "[RU] Обновить промокод (админ) / [EN] Update promocode (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID промокода / [EN] Promocode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "[RU] Данные для обновления / [EN] Update data",
                        "name": "promocode",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.UpdatePromocodeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Promocode"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "[RU] Удаление промокода из базы данных администратором. / [EN] Deletion of a promocode from the database by admin.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-promocodes"
                ],
                "summary": "[RU] Удалить промокод (админ) / [EN] Delete promocode (admin)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "[RU] ID промокода / [EN] Promocode ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.DiscountType": {
            "type": "string",
            "enum": [
                "percent",
                "fixed"
            ],
            "x-enum-varnames": [
                "DiscountTypePercent",
                "DiscountTypeFixed"
            ]
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.Promocode": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discount_type": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.DiscountType"
                },
                "discount_value": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_uses": {
                    "type": "integer"
                },
                "min_amount": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "used_count": {
                    "type": "integer"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
        "github_com_bekzat-kamen_booking_system_api_internal_model.PromocodeDetail": {
            "type": "object",
            "properties": {
                "promocode": {
                    "$ref": "#/definitions/github_com_bekzat-kamen_booking_system_api_internal_model.Promocode"
                },
                "usage_logs": {
                    "type": "object",
                    "additionalProperties": true
                }
            }
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.PromocodesStats": {
            "type": "object",
            "properties": {
                "active_promocodes": {
                    "type": "integer"
                },
                "total_promocodes": {
                    "type": "integer"
                },
                "total_uses": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_bekzat-kamen_booking_system_api_internal_model.UpdatePromocodeRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "discount_value": {
                    "type": "number",
                    "minimum": 0
                },
                "is_active": {
                    "type": "boolean"
                },
                "max_uses": {
                    "type": "integer",
                    "minimum": 0
                },
                "min_amount": {
                    "type": "number",
                    "minimum": 0
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}